[TO REMOVE WHEN FIXED] Poly hack for WinChrome
Chrome on Windows has a problem when rendering some Google Webfonts
(LOLILOL GOOGLE).
The fix is to load a SVG font before the .woff/.ttf. The problem is
that Poly does not come with a SVG variant. I converted the ttf one to
SVG using the FontSquirrel converter.

However, I don't want to serve an automatically converted font to
browsers that work correctly. So I did an ugly hack in javascript to
change the font on the body ONLY on Chrome and Windows.

The problem is that it adds a micro delay when loading the page. It's
visible, it's bad, but meanwhile a fix by the Chrome team, I think that
it's enough.

When the fix will be released, this commit has to be removed. And
ponies will arose and save the world from the darkness of bad
typography. An age of peace and beauty will then be.

Thanks Chrome for Windows. At least, on IE, I can target it.